+++ etl/processor/processor_test.go
@@ -68,7 +68,7 @@ func TestNewProcessor(t *testing.T) {
 			MainSection: &types.MainSection{
 				Start: types.DropStart{
 					PulseData: types.Pulse{
-						PulseNo:        int(gen.PulseNumber()),
+						PulseNo:        int64(gen.PulseNumber()),
 						NextPulseDelta: 10,
 						PrevPulseDelta: 10,
 					},
@@ -104,8 +104,8 @@ func TestProcessor_process_EmptyPrev(t *testing.T) {
 	sm := mock.NewStorageSetterMock(t)
 	sm.SavePulseMock.Set(func(pulse models.Pulse) (err error) {
 		require.Equal(t, jd.MainSection.Start.PulseData.PulseNo, pulse.PulseNumber)
-		require.Equal(t, jd.MainSection.Start.PulseData.PrevPulseDelta, pulse.PulseNumber-pulse.PrevPulseNumber)
-		require.Equal(t, jd.MainSection.Start.PulseData.NextPulseDelta, pulse.NextPulseNumber-pulse.PulseNumber)
+		require.Equal(t, int64(jd.MainSection.Start.PulseData.PrevPulseDelta), pulse.PulseNumber-pulse.PrevPulseNumber)
+		require.Equal(t, int64(jd.MainSection.Start.PulseData.NextPulseDelta), pulse.NextPulseNumber-pulse.PulseNumber)
 		return nil
 	})
 	sm.SaveJetDropDataMock.Set(func(jetDrop models.JetDrop, records []models.Record) (err error) {
@@ -144,8 +144,8 @@ func TestProcessor_process_SeveralPrev(t *testing.T) {
 	sm := mock.NewStorageSetterMock(t)
 	sm.SavePulseMock.Set(func(pulse models.Pulse) (err error) {
 		require.Equal(t, jd.MainSection.Start.PulseData.PulseNo, pulse.PulseNumber)
-		require.Equal(t, jd.MainSection.Start.PulseData.PrevPulseDelta, pulse.PulseNumber-pulse.PrevPulseNumber)
-		require.Equal(t, jd.MainSection.Start.PulseData.NextPulseDelta, pulse.NextPulseNumber-pulse.PulseNumber)
+		require.Equal(t, int64(jd.MainSection.Start.PulseData.PrevPulseDelta), pulse.PulseNumber-pulse.PrevPulseNumber)
+		require.Equal(t, int64(jd.MainSection.Start.PulseData.NextPulseDelta), pulse.NextPulseNumber-pulse.PulseNumber)
 		return nil
 	})
 	sm.SaveJetDropDataMock.Set(func(jetDrop models.JetDrop, records []models.Record) (err error) {
